Implement PensjonTransaksjonId interface in relevant request classes and refactor imports for consistency #deploy-test-dolly-backend
