+++ apps/dolly-backend/src/main/java/no/nav/dolly/bestilling/pensjonforvalter/domain/PensjonTransaksjonId.java
@@ -1,0 +1,4 @@
+package no.nav.dolly.bestilling.pensjonforvalter.domain;
+
+public interface PensjonTransaksjonId {
+}
