+++ apps/dolly-backend/src/main/java/no/nav/dolly/bestilling/pensjonforvalter/utils/PensjonforvalterHelper.java
@@ -4,9 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import no.nav.dolly.bestilling.pensjonforvalter.PensjonforvalterConsumer;
 import no.nav.dolly.bestilling.pensjonforvalter.domain.AlderspensjonVedtakDTO;
-import no.nav.dolly.bestilling.pensjonforvalter.domain.AlderspensjonVedtakRequest;
+import no.nav.dolly.bestilling.pensjonforvalter.domain.PensjonTransaksjonId;
 import no.nav.dolly.bestilling.pensjonforvalter.domain.PensjonforvalterResponse;
 import no.nav.dolly.domain.jpa.TransaksjonMapping;
 import no.nav.dolly.domain.resultset.SystemTyper;
@@ -14,7 +13,6 @@ import no.nav.dolly.errorhandling.ErrorStatusDecoder;
 import no.nav.dolly.service.TransaksjonMappingService;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.time.LocalDateTime;
@@ -22,9 +20,7 @@ import java.util.Comparator;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static no.nav.dolly.bestilling.pensjonforvalter.utils.PensjonforvalterUtils.basicAlderspensjonRequest;
 import static no.nav.dolly.bestilling.pensjonforvalter.utils.PensjonforvalterUtils.basicAlderspensjonRequestDTO;
-import static no.nav.dolly.domain.resultset.SystemTyper.PEN_AP;
 import static no.nav.dolly.errorhandling.ErrorStatusDecoder.encodeStatus;
 import static org.apache.poi.util.StringUtil.isNotBlank;
 
@@ -35,12 +31,11 @@ public class PensjonforvalterHelper {
 
     private final ErrorStatusDecoder errorStatusDecoder;
     private final ObjectMapper objectMapper;
-    private final PensjonforvalterConsumer pensjonforvalterConsumer;
     private final TransaksjonMappingService transaksjonMappingService;
 
     @SuppressWarnings("java:S3740")
     public Mono<TransaksjonMapping> saveAPTransaksjonId(String ident, String miljoe, Long bestillingId,
-                                                        SystemTyper type, Object vedtak) {
+                                                        SystemTyper type, PensjonTransaksjonId vedtak) {
 
         log.info("Lagrer transaksjon for {} i {} ", ident, miljoe);
 
@@ -49,7 +44,7 @@ public class PensjonforvalterHelper {
     }
 
     public Mono<TransaksjonMapping> saveTransaksjonId(String ident, String miljoe, Long bestillingId,
-                                                      SystemTyper type, Object vedtak) {
+                                                      SystemTyper type, PensjonTransaksjonId vedtak) {
 
         return transaksjonMappingService.delete(ident, miljoe, type.name())
                 .then(transaksjonMappingService.save(
@@ -100,19 +95,6 @@ public class PensjonforvalterHelper {
         } else {
             return errorStatusDecoder.getErrorText(HttpStatus.valueOf(httpStatus.getStatus()), httpStatus.getReasonPhrase());
         }
-    }
-
-    public Flux<AlderspensjonVedtakRequest> hentTransaksjonMappingAP(String ident, String miljoe) {
-
-        return transaksjonMappingService.getTransaksjonMapping(PEN_AP.name(), ident, miljoe)
-                .map(mapping -> {
-                    try {
-                        return objectMapper.readValue(mapping.getTransaksjonId(), AlderspensjonVedtakRequest.class);
-                    } catch (JsonProcessingException e) {
-                        log.error("Feil ved deserialisering av transaksjonId", e);
-                        return basicAlderspensjonRequest(ident, Set.of(miljoe));
-                    }
-                });
     }
 
     public Mono<AlderspensjonVedtakDTO> hentSisteVedtakAP(String ident, String miljoe) {
